+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Added: `Pointer.arrayGet()` and `Pointer.arraySet()`.
+* Added: `Callfunc.newCallback()`.
 
 ## 0.1.0 (2019-06-19)